+++ src/Order/Changelog.md
@@ -7,7 +7,8 @@
 
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support
-- Added the `OrderStateChangedToPending` event
+- Added the `WITHDRAWN` order status
+- Added the `OrderStateChangedToPending` and `OrderWasWithdrawn` events
 
 ## 5.x Series